Added JWT authentication, need to fix ts errors

+++ mern/backend/controllers/skillControler.ts
@@ -3,25 +3,27 @@ import asyncHandler from 'express-async-handler';
 
 // Models
 import Skill from '../models/skillModel';
+import User from '../models/userModel';
 
 export const getSkills = asyncHandler(async (req: Request, res: Response) => {
-  const skills = await Skill.find();
+  const skills = await Skill.find({ user: req.user.id });
   res.status(200).json(skills);
 });
 
 export const createSkill = asyncHandler(async (req: Request, res: Response) => {
-  console.log(req.body)
+  console.log(req.body);
   if (Object.keys(req.body).length === 0) {
     res.status(400);
     throw new Error('Please add a skill field');
   }
 
   const skill = await Skill.create({
+    user: req.user.id,
     name: req.body.name,
     level: req.body.level,
     desc: req.body.desc,
-    inProgress: req.body.inProgress
-  })
+    inProgress: req.body.inProgress,
+  });
 
   res.status(200).json(skill);
 });
@@ -34,9 +36,24 @@ export const updateSkills = asyncHandler(
       throw new Error('Skill not found');
     }
 
-    const updatedSkill = await Skill.findByIdAndUpdate(req.params.id, req.body, {
-      new: true
-    })
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error('User not found');
+    }
+
+    if (skill.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error('User not authorized');
+    }
+
+    const updatedSkill = await Skill.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        new: true,
+      }
+    );
 
     res.status(200).json(updatedSkill);
   }
@@ -50,7 +67,18 @@ export const deleteSkills = asyncHandler(
       throw new Error('Skill not found');
     }
 
-    await Skill.findByIdAndDelete(req.params.id, req.body)
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error('User not found');
+    }
+
+    if (skill.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error('User not authorized');
+    }
+
+    await Skill.findByIdAndDelete(req.params.id, req.body);
     res.status(200).json({ message: `Skill ${req.params.id} deleted` });
   }
 );
